migration version runnner now supports interleaved migrations

+++ src/FluentMigrator.Tests/Integration/Migrations/Interleaved/Pass3/1_User.cs
@@ -6,7 +6,7 @@
         public override void Up()
         {
             Create.Table("User")
-                .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+                .WithColumn("Id").AsInt32().PrimaryKey().Identity()
                 .WithColumn("Name").AsString();
         }
